rag-daytona.v2/requirements.txt
# FastAPI and server
fastapi>=0.115.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Pydantic for data validation
pydantic>=2.5.0

# Gemini API client
google-genai>=1.33.0
google-generativeai>=0.8.0

# Google Custom Search API (for web search)
google-api-python-client>=2.100.0

# Vector search (CPU version - lighter than GPU)
faiss-cpu>=1.7.4

# Scientific computing
numpy==1.24.3

# Redis client
redis[asyncio]>=5.0.0

# HTTP client for health checks
httpx>=0.24.0

# Environment variables
python-dotenv==1.0.1

# ML/Embeddings (HEAVY but required for service to start)
sentence-transformers>=2.2.0
langchain-huggingface>=0.0.1
huggingface-hub>=0.16.0
torch>=2.0.0


# Qdrant Client (for Vector DB)
qdrant-client>=1.7.0

# LLM Providers
openai>=1.12.0
anthropic>=0.18.0
requests>=2.31.0
ollama>=0.1.0

# LLM Providers
openai>=1.12.0
anthropic>=0.18.0
requests>=2.31.0
ollama>=0.1.0
groq>=0.4.2

# Optimization (ONNX Runtime)
optimum[onnxruntime]>=1.16.0
safetensors>=0.4.0
